Added points to StudentDto, points now is Integer, added Postgres and…

+++ src/main/java/ru/t1murcoder/controller/dto/StudentDto.java
@@ -14,6 +14,5 @@ public class StudentDto {
     private String name;
     private String surname;
     private String username;
-    private Float points;
-    //TODO: сделать посещения
+    private Integer points;
 }
